source/requirements.txt
torch
pytorch-lightning
hydra-core
omegaconf
transformers
torchmetrics
wandb
numpy
pandas